implement changes

+++ src/main/java/com/perennial/doctorappointmentbooking/repo/PaymentRepo.java
@@ -1,10 +1,9 @@
 package com.perennial.doctorappointmentbooking.repo;
-
 import com.perennial.doctorappointmentbooking.entity.Patient;
 import com.perennial.doctorappointmentbooking.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepo extends JpaRepository<Payment,Integer> {
-    Payment findByPaymentId(Integer payment_id);
+public interface PaymentRepo extends JpaRepository<Payment,Long> {
+    Payment findByPaymentId(int payment_id);
 
 }
